Add ELF text segment extracter
elf2int.sh prints the text segment in the form expected by the decoder.

Simple, not entirely stable cross-assembler:

	llvm-mc -assemble -arch aarch64 -filetype obj -o foo.o foo.s

For more real code, you need to cross-compile, for example with Go,
where it is extremely simple compared to the C toolchains. And because
it's statically linked, there's lots of code!

	GOARCH=arm64 go build -o foo