farmdec: prepare header, move out main, add C++ example
The header file uses a few tricks. For example, it defines the
convenient integer typedefs only for internal use. External users
only see the equivalent stdint.h types, hence avoiding pollution.

Furthermore, in lieu of a common prefix for the enums and all of their
values, there's a farmdec namespace for C++. Granted, in C it's all
exposed, but _if_ there _were_ a C program that wants to link against
libfarmdec, it's unlikely to have conflicting enum names and values
and more likely to use the enums as they are.

I'll add a common prefix for the exposed functions, however.