movn,movz: combine into A64_MOV_IMM
The virtual instruction MOV_IMM ought to include *all* moves whose
value can be precalculated; there's no reason to differentiate
between one of the immediate MOV aliases and MOVZ/MOVN because the
for these the full 32-bit or 64-bit value can be calculated as well.
Only MOVK is different and needs to keep bits intact.