simd: add Ops, flags, decoding to kind
This took a bunch of effort, and I procrastinated it.

The manual splits scalar and vector instructions, but they share so
much I meticulously pieced them together in eight pages of handwritten
decoding notes.

I'm a bit unhappy with the consistency of the opcode naming, and
especially with the split between vector instructions that have the
"Scalar" flag and scalar FP instructions that have different Ops and
don't use the Scalar flag. However, this is not significant enough
to warrant refactoring.

Frequency statistics on the SPEC binaries:

	SIMD/Copy:        786 0.257283
	SIMD/ThreeSame:   734 0.240262
	SIMD/Extract:     480 0.157119
	SIMD/ModifiedImm: 418 0.136825
	SIMD/TwoRegMisc:  393 0.128642
	SIMD/Reduce:      212 0.0693944
	SIMD/ShiftByImm:   24 0.00785597
	SIMD/TableLookup:   8 0.00261866